+++ packages/notifi-react-card/lib/components/intercom/index.ts
@@ -1,4 +1,4 @@
-import '../common/defaults.css';
+import '../defaults.css';
 
 export * from './ChatMessageDate';
 export * from './ChatMessageSection';
