+++ packages/notifi-react-card/lib/components/subscription/index.ts
@@ -1,4 +1,4 @@
-import '../common/defaults.css';
+import '../defaults.css';
 
 export * from './EventTypeBroadcastRow';
 export * from './EventTypeDirectPushRow';
